refactor(autoware_pid_longitudinal_controller): move Node functions into pid_longitudinal_controller_node.cpp; drop unused includes
Move create_config() and all PidLongitudinalControllerNode method
implementations (ctor, paramCallback, isReady, run, emitLogs,
publishDebugData, publishVirtualWallMarker, setupDiagnosticUpdater,
checkControlState) out of pid_longitudinal_controller.cpp into a new
pid_longitudinal_controller_node.cpp, matching the extract-core-logic
naming convention of <package>.cpp for core logic and <package>_node.cpp
for the ROS 2 node wrapper.

Also remove autoware_utils/ros/marker_helper.hpp and autoware_utils/geometry/
normalization.hpp includes that were unused (pre-existing dead includes,
unrelated to the createStopVirtualWallMarker calls, which come from
autoware/motion_utils/marker/marker_helper.hpp). marker_helper.hpp
happened to pull in visualization_msgs/msg/marker_array.hpp as a side
effect, so switch the marker.hpp include to marker_array.hpp directly
since MarkerArray is the type actually used.
